Organize file struture and style, implement SCSS
