+++ card-draw/src/App.tsx
@@ -1,5 +1,3 @@
-import './index.css';
-import './reset.css';
 import CardDraw from './CardDraw';
 import Footer from './Footer';
