+++ card-draw/src/scss/components/numfield.scss
@@ -1,0 +1,71 @@
+@use '../util/color' as color;
+@use '../util/fonts' as font;
+@use '../util/mixins' as mixin;
+
+.numfield {
+  @include mixin.flex($dir: column, $gap: 4px);
+  &-title {
+    font-size: 1.2em;
+    font-weight: 300;
+    user-select: none;
+  }
+  
+  &-inner {
+    @include mixin.flex($align: space-between, $x-align: center);
+    @include mixin.border();
+    &:hover {
+      border-color: color.$accent;
+    }
+  }
+  
+  &-value {
+    @include font.miso(1.5em, color.$lite, 700);
+    line-height: 0;
+    background-color: color.$darken;
+    border: unset;
+    padding: 7.5px 8px;
+    width: 45px;
+    border-radius: 4px 0 0 4px;
+    &:hover {
+      background-color: color.$accent-dark;
+    }
+    &:focus {
+      outline: unset;
+    }  
+  }
+
+  &-controls {
+    @include mixin.flex($dir: column, $align: normal);
+    & > * {
+      padding: 3px 8px;
+      border: none;
+      background-color: color.$dark-lite;
+      cursor: pointer;
+      &:hover {
+        background-color: color.$accent;
+      }
+    }
+  }
+
+  &-increment {
+    border-radius: 0 2px 0 0;
+  }
+  
+  &-decrement {
+    border-radius: 0 0 2px 0;
+  }
+}
+
+.caret {
+  fill: color.$lite;
+  width: 12px; 
+  &-increment {
+    @extend .caret;
+  }
+  &-decrement {
+    @extend .caret;
+    position: relative;
+    transform: rotate(180deg);
+    top: 1px;
+  }
+}
