+++ card-draw/src/scss/util/mixins.scss
@@ -1,0 +1,14 @@
+@use '../util/color' as color;
+
+@mixin flex($dir: row, $wrap: nowrap, $align: center, $x-align: stretch, $gap: none) {
+  display: flex;
+  flex-flow: $dir $wrap;
+  justify-content: $align;
+  align-items: $x-align;
+  gap: $gap;
+}
+
+@mixin border($radius: 4px, $width: 2px, $color: color.$lite-dark) {
+  border-radius: $radius;
+  border: $width solid $color;
+}
